refactor: Rename file `tests/helpers.rs` to `tests/test_helpers.rs`
